Separate LocalModelCache webhook from KServe controller. (#4941)
