+++ config/certmanager/localmodel/certificate.yaml
@@ -1,0 +1,16 @@
+# The following manifests contain a self-signed issuer CR and a certificate CR.
+# More document can be found at https://docs.cert-manager.io
+apiVersion: cert-manager.io/v1
+kind: Certificate
+metadata:
+  name: localmodel-serving-cert  # this name should match the one appeared in kustomizeconfig.yaml
+  namespace: kserve
+spec:
+  # $(SERVICE_NAME) and $(SERVICE_NAMESPACE) will be substituted by kustomize
+  commonName: $(webhookServiceName).$(kserveNamespace).svc
+  dnsNames:
+    - $(webhookServiceName).$(kserveNamespace).svc
+  issuerRef:
+    kind: Issuer
+    name: selfsigned-issuer
+  secretName: localmodel-webhook-server-cert # this secret will not be prefixed, since it's not managed by kustomize
