+++ config/localmodels/kustomization.yaml
@@ -1,5 +1,66 @@
 apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
 
+namespace: kserve
+
 resources:
 - manager.yaml
+- ../webhook/localmodel
+- ../certmanager/localmodel
+
+patches:
+- path: webhook_cainjection_patch.yaml
+
+replacements:
+# Replace the namespace with the namespace of the controller manager.
+- source:
+    fieldPath: metadata.namespace
+    kind: Deployment
+    name: kserve-localmodel-controller-manager
+    version: v1
+  targets:
+  - fieldPaths:
+    - webhooks.*.clientConfig.service.namespace
+    select:
+      kind: ValidatingWebhookConfiguration
+      name: localmodelcache.serving.kserve.io
+  - fieldPaths:
+    - metadata.annotations.[cert-manager.io/inject-ca-from]
+    options:
+      delimiter: '/'
+      index: 0
+    select:
+      kind: ValidatingWebhookConfiguration
+      name: localmodelcache.serving.kserve.io
+  - fieldPaths:
+    - spec.commonName
+    - spec.dnsNames.0
+    options:
+      delimiter: '.'
+      index: 1
+    select:
+      kind: Certificate
+      name: localmodel-serving-cert
+      namespace: kserve
+# Replace localmodel webhook service name
+- source:
+    fieldPath: metadata.name
+    kind: Service
+    name: localmodel-webhook-server-service
+    version: v1
+  targets:
+  - fieldPaths:
+    - webhooks.*.clientConfig.service.name
+    select:
+      kind: ValidatingWebhookConfiguration
+      name: localmodelcache.serving.kserve.io
+  - fieldPaths:
+    - spec.commonName
+    - spec.dnsNames.0
+    options:
+      delimiter: '.'
+      index: 0
+    select:
+      kind: Certificate
+      name: localmodel-serving-cert
+      namespace: kserve
