+++ config/webhook/localmodel/service.yaml
@@ -1,0 +1,11 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: localmodel-webhook-server-service
+  namespace: kserve
+spec:
+  ports:
+    - port: 443
+      targetPort: webhook-server
+  selector:
+    control-plane: kserve-localmodel-controller-manager
