+++ config/webhook/manifests.yaml
@@ -182,30 +182,3 @@ webhooks:
           - UPDATE
         resources:
           - servingruntimes
----
-apiVersion: admissionregistration.k8s.io/v1
-kind: ValidatingWebhookConfiguration
-metadata:
-  creationTimestamp: null
-  name: localmodelcache.serving.kserve.io
-webhooks:
-  - clientConfig:
-      service:
-        name: $(webhookServiceName)
-        namespace: $(kserveNamespace)
-        path: /validate-serving-kserve-io-v1alpha1-localmodelcache
-    failurePolicy: Fail
-    name: localmodelcache.kserve-webhook-server.validator
-    sideEffects: None
-    admissionReviewVersions: ["v1beta1"]
-    rules:
-      - apiGroups:
-          - serving.kserve.io
-        apiVersions:
-          - v1alpha1
-        operations:
-          - CREATE
-          - UPDATE
-          - DELETE
-        resources:
-          - localmodelcaches
