+++ hack/setup/quick-install/kserve-standard-mode-full-install-with-manifests.sh
@@ -76934,6 +76934,18 @@ spec:
   selector:
     control-plane: llmisvc-controller-manager
 ---
+apiVersion: v1
+kind: Service
+metadata:
+  name: localmodel-webhook-server-service
+  namespace: kserve
+spec:
+  ports:
+  - port: 443
+    targetPort: webhook-server
+  selector:
+    control-plane: kserve-localmodel-controller-manager
+---
 apiVersion: apps/v1
 kind: Deployment
 metadata:
@@ -77075,7 +77087,26 @@ spec:
               fieldPath: metadata.namespace
         image: kserve/kserve-localmodel-controller:latest
         imagePullPolicy: Always
+        livenessProbe:
+          failureThreshold: 5
+          httpGet:
+            path: /healthz
+            port: 8081
+          initialDelaySeconds: 30
+          timeoutSeconds: 5
         name: manager
+        ports:
+        - containerPort: 9443
+          name: webhook-server
+          protocol: TCP
+        readinessProbe:
+          failureThreshold: 5
+          httpGet:
+            path: /readyz
+            port: 8081
+          initialDelaySeconds: 30
+          periodSeconds: 5
+          timeoutSeconds: 5
         resources:
           limits:
             cpu: 100m
@@ -77091,12 +77122,21 @@ spec:
           privileged: false
           readOnlyRootFilesystem: true
           runAsNonRoot: true
+        volumeMounts:
+        - mountPath: /tmp/k8s-webhook-server/serving-certs
+          name: cert
+          readOnly: true
       securityContext:
         runAsNonRoot: true
         seccompProfile:
           type: RuntimeDefault
       serviceAccountName: kserve-localmodel-controller-manager
       terminationGracePeriodSeconds: 10
+      volumes:
+      - name: cert
+        secret:
+          defaultMode: 420
+          secretName: localmodel-webhook-server-cert
 ---
 apiVersion: apps/v1
 kind: Deployment
@@ -77278,6 +77318,20 @@ spec:
     kind: Issuer
     name: selfsigned-issuer
   secretName: llmisvc-webhook-server-cert
+---
+apiVersion: cert-manager.io/v1
+kind: Certificate
+metadata:
+  name: localmodel-serving-cert
+  namespace: kserve
+spec:
+  commonName: localmodel-webhook-server-service.kserve.svc
+  dnsNames:
+  - localmodel-webhook-server-service.kserve.svc
+  issuerRef:
+    kind: Issuer
+    name: selfsigned-issuer
+  secretName: localmodel-webhook-server-cert
 ---
 apiVersion: cert-manager.io/v1
 kind: Certificate
@@ -78883,7 +78937,7 @@ apiVersion: admissionregistration.k8s.io/v1
 kind: ValidatingWebhookConfiguration
 metadata:
   annotations:
-    cert-manager.io/inject-ca-from: kserve/serving-cert
+    cert-manager.io/inject-ca-from: kserve/localmodel-serving-cert
   creationTimestamp: null
   name: localmodelcache.serving.kserve.io
 webhooks:
@@ -78891,7 +78945,7 @@ webhooks:
   - v1beta1
   clientConfig:
     service:
-      name: kserve-webhook-server-service
+      name: localmodel-webhook-server-service
       namespace: kserve
       path: /validate-serving-kserve-io-v1alpha1-localmodelcache
   failurePolicy: Fail
