+++ pkg/controller/v1alpha1/localmodel/webhook/localmodelcache_validator.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package localmodelcache
+package webhook
 
 import (
 	"context"
